backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
boto3==1.29.7
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
jinja2==3.1.2
premailer==3.10.0

